requirements.txt
pymongo==3.12.0
beautifulsoup4==4.9.3
requests==2.26.0
kafka-python==2.0.2